requirements.txt
Flask
Flask-SocketIO
python-dotenv
pywebpush
pillow
Flask-Limiter
Flask-Moment
Flask-Session
git+https://github.com/chronosirius/nosql-db-client
simple-websocket
marko